Add RequestsResponseSchema base class
Base class for schemas that validate `requests`-library responses. This
will avoid having duplicate code in different schemas, or in the code
that calls those schemas, for example to call `response.json()` and
handle errors when the response isn't valid JSON, or if it is valid to
pass the params to marshmallow for validation, etc.