Fix module item configuration error page
By default webargs searches the request's query string, form data and
JSON body for valid parameters. In the case of
`ConfigureModuleItemSchema` the parameters are in fact in the form body.
If the form body is, for example, missing a required parameter then
webargs:

1. Tries to validate the request based on its query string, which fails
   because there's no query string

2. Tries to validate the request based on its form body, which fails
   because there's one required parameter missing

3. Tries to validate the request based on its JSON body, which fails
   because the body isn't JSON

4. Finally gives up and raises a validation error

But the error message is based on the final, JSON attempt, so the user
sees:

    {'json': ['Invalid JSON body.']}

Fix this by allowing schema classes to override webargs's default
locations by having a class attribute `locations` (list of strings), and
add `locations = ["form"]` to `ConfigureModuleItemSchema`.

The user now sees the desired error message:

    Hypothesis received an invalid request. There were problems with these request parameters:

    * foo:
      * Missing data for required field.

See https://webargs.readthedocs.io/en/latest/quickstart.html#request-locations