Add CanvasAPIAccessTokenError
Define a new CanvasAPIAccessTokenError and raise it whenever we get a
401 (meaning invalid, expired or deleted access token) response from
Canvas, or if we don't have an access token for the user at all.

This will enable the views, in future commits, to distinguish between
CanvasAPIAccessTokenError and CanvasAPIServerError and send back
different kinds of responses so that the frontend can also distinguish
them.