Send 400 Bad Request responses for Canvas API failures
When our frontend code calls our list_files or via_url proxy API and the
Canvas API request fails or can't be made for any reason, send a JSON
400 Bad Request response back to our frontend code.

If the failure is due to a missing, expired or deleted access token the
response contains no error message or details, because we don't want the
frontend to show any error to the user:

    400 Bad Request
    {"error_message": null, "details": null}

If the failure is due to an unsuccessful response from Canvas then we do
pass an error message and details to the frontend, for showing to the
user. These include the status code and JSON response body from Canvas:

    400 Bad Request
    {
      "error_message": "Calling the Canvas API failed",
      "details": {
        "status_code": 404,
        "json_body": {"errors": [{"message": "The specified resource does not exist."}]}
      }
    }

In cases where the failure does not include a response from Canvas with
a status code or JSON body, for example a network error that meant that
no response was ever received, or a 500 Server Error from Canvas with no
JSON body, the details are null. We should improve this to provide some
details of what went wrong:

    400 Bad Request
    {
      "error_message": "Calling the Canvas API failed",
      "details": {"status_code": null, "json_body": null}
    }