Catch and report 4xx errors from API calls to the backend
Check for error messages from the backend and translate them to
`ApiError` exceptions in the frontend which UI components can test for.

After discussion with @seanh, I understood that the server does not
explicitly indicate when authorization with Canvas fails. Instead it
returns a generic response with no `error_message` property and expects
the frontend to use its "default" handling, which in this case means
showing the Authorization prompt.

 - Map 4xx and 5xx error responses from server to `ApiError` exceptions
   locally
 - Remove `AuthorizationError` exception