Catch and display API errors in LMS file picker dialog
Catch API errors that occur in calls to the backend to list LMS files
and when an error happens, display either an authorization prompt or
error details in the file picker depending on the error.

If an `ApiError` exception was thrown and no error details are provided,
we assume the problem is failed authorization and an "Authorize"
button is shown. If the `ApiError` included error details or some other type
of error happened then we show those details together with a "Try again"
button. The "Try again" button performs exactly the same authorization
flow as the "Authorize" button.

In the implementation, the `LMSFilePicker` component's state
representation has been changed to eliminate invalid states (such as eg.
being in an "authorizing" and "loading" state at the same time) and make
it clearer that the file picker can be in one of 4 main states
(authorizing, fetching data, displaying the file list, displaying an
error).