Change _access_token() to _oauth2_token()
Change the _access_token() helper to _oauth2_token() and have it return
the whole OAuth2Token model instead of just the access_token string.

We're going to be needing the refresh_token string from the OAuth2Token
in the future, not just the access_token string, so this change will be
needed.