Always attempt to fetch files when the LMS file picker is shown
Instead of expecting the server to provide a hint as to whether the user
has authorized the H LMS app to access their LMS files, just try to
fetch files when the file picker is shown and display a prompt to
authorize if that fails.

As a result, the user is no longer prompted to authorize if they have
already authorized the LMS in a previous session.