Implement content URL fetching at assignment launch for Canvas files
Implement a small client-side application which handles fetching the
temporary Via URL for assignment launches involving PDF files in Canvas.

Initially the app renders a loading indicator and makes a request for
the assignment content URL. The backend will then retrieve that URL by
making API calls to Canvas. When the URL is returned, the client app
renders an iframe with that URL.

If the URL cannot be fetched because authorization is required, the
client will prompt the user for authorization. If an error occurs, that
will be displayed to the user.

---

The implementation reuses many of the components from the file picker
app, so it is implemented as a new component in the file_picker_v2
bundle.