Remove focus outline around table when not using keyboard input
Import the `outline-on-keyboard-focus` mixin from the client and
focus-visible polyfill and use it to remove the focus outline from the
file list when the user has focused the table via non-keyboard
navigation.

Ideally the focus indicator would be drawn around the table wrapper, but
there is no `:focus-within` equivalent for `:focus-visible` yet.