Render Hypothesis client config into page if set but empty
When provisioning is disabled, `context.hypothesis_config` is set to an
empty dictionary. Because this is a falsey value in Python, the config
was not rendered into the page, causing the client-side JS to fail.

Modify the base template to check for the existence of JS config related
values instead of falseyness.

The client continued to work when using the legacy (non-`new_oauth`)
mode because the new_lti_launch.html.jinja2 template had duplicated
logic to render this config into the page.

Fixes #755