Work around bug in PyLTI's handling of URL params with percent coded chars
Fix an OAuth 1.0 signature verification failure when handling an LTI
launch request where:

 - There is a "url" query param (aka. the launch request is "URL configured")
 - The "url" query param, _after decoding_, contains percent-encoded
   query parameters.

For example, given an LTI launch request with the following URL:

  "https://lms.hypothes.is/lti_launches?url=https%3A%2F%2Fen.wikipedia.org%2Fwiki%2FG%25C3%25B6reme_National_park"

The LMS would generate an OAuth signature base string using the "url"
parameter "https://en.wikipedia.org/wiki/G%C3%B6reme_National_ark".

When our lms app received the request, it would parse the incoming
request so `request.params["url"]` was "https://en.wikipedia.org/wiki/G%C3%B6reme_National_ark".

However when PyLTI generates the signature base string for comparison,
it _decoded_ this value and used:

  "https://en.wikipedia.org/wiki/Göreme_National_Park"

The result is that the signatures did not match and the request aborted
with a 403 HTTP status.

Since this problem only affects parameters that PyLTI extracts from the
URL and not those passed to `verify_request_common` in the `parameters`
kwarg, and query params are present in both, the workaround here is to
exclude the query string + fragment from the value passed to
`verify_request_common` as the `url` argument.

Fixes #689