Fix file icons for selected files displaying above sticky header
File icons for selected files have a CSS `filter` effect applied. This
creates a new stacking context which caused them to be painted on top of
the `position: sticky` table headers when the content was scrolled.

Fix this by applying a z-index to the table header.

See also https://github.com/hypothesis/lms/issues/759