Add an API endpoint to record a grading LTI launch URL in SpeedGrader
Add an API endpoint that the frontend will use to record a
grading-oriented LTI launch URL in Canvas's SpeedGrader for use when a teacher
grades the current user's annotations on an assignment.

This operation will be triggered by the frontend rather than by the
backend during the /lti_launches request, where all the necessary
information is already available, because it involves making (possibly
slow) calls to the external LMS. Instead these potentially slow
operations can happen concurrently with the assignment loading in the
user's browser.

 - Add `/api/lti/submissions` endpoint to record SpeedGrader Launch URL
 - Add schema to validate JSON body of requests to this endpoint