Show helpful error if file auth fails
If the authentication fails show a more detailed error message about why
the request failed so hopefully the user can get out of the infinite
allow file access loop. This is a fix for
https://github.com/hypothesis/lms/issues/767.

Add lmsRegisteredUrl which is the url that the lms application instance was
registered with (aka the url in the database). This is different than lmsUrl
which is the url of the lms from the client perspective (aka in the url
bar).