Work around SpeedGrader LTI launch URL encoding bug
Query params in LTI launch URLs submitted to Canvas for use in SpeedGrader get
doubly-encoded so they are still percent-encoded after Pyramid/WSGI have parsed
the request. Implement a workaround by decoding the query param a second time in
the LTI Launch view and signpost the potential issue when generating the
SpeedGrader launch URL.

More generic solutions are possible, but they could have unexpected interactions
with different parts of the system. Therefore this PR just handles the one query
param where this is a problem.