Use Hypothesis client's "focus mode" to focus on a user in SpeedGrader
In Canvas's SpeedGrader view, configure the Hypothesis client to focus
on a specific user using the "focus" config [1]. This requires making a
call to the Hypothesis API to look up the user's display name first.

This change depends on [2] being deployed first to implement the
`GET /api/users/{userid}` endpoint in h.

[1] https://github.com/hypothesis/client/pull/1291
[2] https://github.com/hypothesis/h/pull/5672