Extract `GET /api/users/{userid}` call into service
Following the discussion in [1] create a service that provides a
high-level client for the "h" API and move the logic to fetch users
there instead of accessing the "hapi" service directly from within the
LTI launch views.

Also create an `HUser` type to represent h users generated from LTI
users and return an instance of that object from the high-level API
client.

[1] https://github.com/hypothesis/lms/pull/888#discussion_r313453460