Run gunicorn and gulp in one terminal
Change `make dev` to run both gunicorn (development web server) and
`gulp watch` (frontend assets build) in one terminal, using honcho to
interleave the stdouts of both processes into the one terminal. (honcho
is added as a new dev dependency.)

The end result is that you still run `make dev` as before but it'll now
run both gunicorn and gulp. You no longer need to run gulp separately
yourself.

* `make web` is a new command that now does what `make dev' used to do:
  runs the web server only

* `make assets` is a new command that wraps `gulp watch`

* Procfile is honcho's config file, which configures `honcho start` to
  run `make web` and `make assets` simultaneously in one terminal.

  Procfile calls Makefile to run `make web` and `make assets`. This
  avoids duplicating the actual `tox` and `gulp` commands that
  underly `make web` and `make assets`. Makefile remains the single
  source of truth for these underlying commands.

* `make dev` is now a wrapper for `honcho start`: it runs `make web` and
  `make assets` simultaneously

Debugging
---------

Unfortunately honcho does mean that, when running `make dev`, neither
the gunicorn or the gulp process is connected to your terminal's stdin:
honcho is. This means that you can't interactively debug gunicorn or
gulp. For example you can't debug the Python code with pdb. Even if
honcho could echo your stdin to the gunicorn process (which it can't) it
wouldn't be much fun to try to interact with pdb in a terminal while
gulp's stdout is being printed to the same terminal.

That's why `make web` and `make assets` are here as separate commands
from `make dev`: if you want to use an interactive debugger with the
gunicorn or gulp process, you should open two terminal tabs or windows
and run `make web` in one and `make assets` in the other. Starting the
processes in this way is less convenient than `make dev` but it means
that each process is running in its own terminal and on its own (no
honcho), so you can debug in peace.

Running only some processes
---------------------------

You can also tell `make dev` which processes to start, instead of
starting all of them:

    $ make dev processes=web
    $ make dev processes=assets
    $ make dev processes='web assets'

This is just an alias for `honcho start [web] [assets]`. This isn't
particularly interesting in an app that only has two honcho processes
but might be more useful in an app with more processes, for example to
run all of the processes bar one (so that you can then run that one
process separately and debug it).

Running honcho directly
-----------------------

Although `make dev` is intended to act as a wrapper, you can of course
also run honcho directly. To have tox install honcho for you and then
run tox's version of honcho via tox directly (bypassing make):

    $ tox -qe py36-dev -- honcho --help

Or to run honcho directly bypassing both tox and make:

    $ .tox/py36-dev/bin/honcho --help

That does require you to have run `make dev` or `make web` or something
previously, so that .tox/py36-dev/ will exist and honcho will have been
installed in it.

If you have installed honcho yourself and its on your $PATH you can
bypass both make and tox and just run honcho directly:

    $ honcho start ...

and this will work with our Procfile.