Combine `LTILaunchResource` h user properties into `h_user`
Combine `LTILaunchResource.h_{username, display_name, userid}`
properties into a single `h_user` property which returns an `HUser`.

This removes the duplication of logic for generating the userid from the
username and authority between `LTILaunchResource` and `HUser`. In
future it will also make it possible to add other convenience properties
to `HUser`.

This is a follow-up to [1]

[1] https://github.com/hypothesis/lms/pull/888#pullrequestreview-274807714