Refactor setState usage to be more explicit
setState does not merge states of keys in a state object. The
previous implementation made it look like this was the case. Refactor
the implementation to make it clear that setState overrides the entire
object's state and use an INITIAL_STATE constant to restore states not
explicitly specified to their initial settings during state changes.