Log "closing a dirty session" as warning
Log the "closing a dirty session" message as a warning (that can be
configured to trigger an alarm in Papertrail) instead of reporting it to
Sentry.

Calling Sentry from this finished callback is a bad idea because it
triggers Sentry data collection code that might access something that
might in turn try to read from the DB, but this finished callback
executes _after_ the pyramid_tm tween has already closed the DB
transaction.

Logging a warning is also consistent with what the same finished
callback in h does, and it removes a call to Sentry which will make
moving this lms app to h-pyramid-sentry easier.