Warn on closing any unclosed DB session
Log a warning if _any_ unclosed DB session has to be closed by our
finished callback, not just if the unclosed session was also dirty. The
finished callback should never have to close a session, if it ever does
this indicates a problem.

This is also consistent with the same warning in h.