Record group info on LTI launch
For analytics purposes we have a need to associate LMS-created h groups
with the school/institution and LMS instance that owns them.

To make this possible, add a new GroupInfo model that associates a
groupid (which is a unique identifier for an h group) with the
corresponding ApplicationInstance and also with various potentially
interesting or useful values from the LTI launch params.

On each LTI launch the GroupInfo for the course's group is either
created or updated using the params from the current launch.

Fixes https://github.com/hypothesis/lms/issues/918