Disable grading UI if evaluation is disabled for assignment
When configuring an assignment in Blackboard there is an "Enable
evaluation" option. When this is _on_, LTI launch requests for
_instructors_ have a `lis_outcome_service_url` param in the request.
When it is _off_ they do not.

As far as I know, the LTI specs do not say anything about whether this
param should be included in requests when instructors launch an
assignment, and the URL is not actually used at all, but it is
nevertheless the only obvious indicator in the LTI launch request of
whether evaluation is enabled.

Fixes #1034