Replace shallow rendering with mocking of imported components
Change UI tests to follow the pattern introduced in
https://github.com/hypothesis/client/pull/1470 of using `mount` rather
than `shallow` Enzyme rendering for all tests, but with imported
components mocked using a `mockImportedComponents` helper.

This makes the tests more consistent, and removes some manual boilerplate
for mocking components.

Some of the `SubmitGradeForm` tests broke due to waiting for async tasks
to complete in a brittle way. This commit resolves that by making the
tests wait for the same loading indicators that the user sees in order
to determine whether network requests have completed or not.