Always use the frontend_app fixture
This frontend_app patch fixture replaces
basic_lti_launch_test.py::frontend_app with a mock, but only in tests
that have the frontend_app fixture as an argument.

Code changes can mean that some other test that was not causing
frontend_app to be accessed now does cause frontend_app to be accessed,
and if those tests don't take frontend_app as an argument then the code
under test will access the real frontend_app not the mock/patched one.

A code change has in fact caused this to happen and as a result the
tests are now crashing--the real frontend_app is being called and is
crashing because a service that it needs isn't registered in these
tests.

Make frontend_app into an autouse fixture so that the mock frontend_app
is always used by all tests in this module.

Generally speaking patch fixtures like frontend_app should always work
like this--we never want unit tests for basic_lti_launch_test.py to be
calling the real frontend_app. That can only lead to those unit tests
crashing or worse behaving differently than we think they are.