Factor out some unused test code
The `if consumer_key != "TEST_OAUTH_CONSUMER_KEY":` was never evaluating
to `True` and so the lines inside the `if` were never being executed and
were decreasing the coverage % of the tests.

Having shared test fixtures that do their own assertions (or in this
case: `raise Exception`!) is a bad pattern anyway. You can't see, by
looking at the tests that use this `ai_getter` fixture, that the tests
are in fact testing that `shared_secret()` was called with the right
value. To see that you'd have to browse to the bottom of the file and
read the `ai_getter` fixture's body code.

Replace this by having each test do its own assertion explicitly. Now
you can actually see what's being tested.