Enable grading in Moodle
Change a bunch of hacky logic about only allowing grading to be enabled
in Blackboard Learn, to hacky logic about only allowing grading to be
enabled in either Blackboard Learn or Moodle.

The code and tests here is hacky and rough because we'll be removing
this logic (and just enabling grading all the time) very soon so it's
not worth any effort.