Send JSON errors for invalid API requests
Send ValidationError's in JSON format when requests to /api/* raise a
ValidationError.

Previously the HTML validation error page was being sent in response to
API requests.

Fix this by forcing the HTML ValidationError exception view to *not*
match requests to /api/* so that the JSON ValidationError exception view
matches instead.