Add launch_presentation_return_url to validation errors
When we raise a ValidationError (e.g. if a required LTI launch param
such as resource_link_id is missing) then if the optional
launch_presentation_return_url param is *also* missing we display the
validation error messages ourselves, since we can't redirect to the
launch_presentation_return_url with the error messages in a query param
as we usually do.

When this happens, also add an error message about the missing or
invalid launch_presentation_return_url to our error message page.