Moving the reports decorator to the view
It only inserts one thing and is only used here.

The only slight contravertial thing is splitting up the tests for
a single thing into separate files. I just didn't want to add
to the tests for the main class, as it's already big.

If we're happy with this split, or some variation of it, I'd suggest
busting up the main test file into smaller easier to grok chunks.