Factor out the _group_info_upsert() helper
Previously the _group_info_upsert() helper was a private helper in the
lti_h service that was responsible for extracting the whitelisted params
(the ones that correspond to columns on the group_info DB table) from
the request, and passing them in to the GroupInfoUpsert service.

Factor this private helper away by:

* Adding a new BaseClass that all SQLAlchemy models inherit from, and
  that provides:

  * iter_properties() and iter_columns() classmethods.
    iter_columns() returns an iterator over all of the model's SQLAlchemy
    columns.

  * An update_from_dict() method that updates the columns of a model
    object from a dict of params.

    Any key in the given dict whose name does not match one of the
    models columns is ignored.

* Having GroupInfoUpsert (now renamed to GroupInfoService) call this new
  update_from_dict() method passing in request.params directly.

  Even though request.params might contain other keys (that don't
  correspond to any of the model's columns) those will be ignored by
  update_from_dict().

  GroupInfoService also removes a couple of specific keys
  ("authority_provided_id" and "id") from the fields to be updated, even
  though those do correspond to columns. We don't want to update those
  columns.

This means that a whitelist of the fields to be updated is no longer
needed. The list is computed dynamically by getting the list of the
model's column names, then subtracting "authority_provided_id" and "id".

As a result, _group_info_upsert() is no longer needed, since that
whitelisting was all it provided,

In addition to the above refactoring, this commit also:

* Renames GroupInfoUpsert to GroupInfoService
* Renames GroupInfoUpsert/GroupInfoService's __call__ method to upsert()
* Refactors an includeme() test and some other tests