Use PyramidRequestSchema to parse Pyramid requests
Use PyramidRequestSchema to parse the Pyramid request, which is what
we're really doing here.

There's not a _huge_ amount of benefit here but we have a schema base
class for parsing and validating Pyramid requests, so let's use it.

PyramidRequestSchema handles parsing the request object, for example
deciding which of request.GET, request.POST, request.params and
request.json_body and others (the URL itself, the headers, cookies, file
uploads) to use (or a combination thereof), and dealing with the fact
that some of these are NestedMultiDicts not dicts.

PyramidRequestSchema is a wrapper for webargs's pyramidparser:

https://github.com/marshmallow-code/webargs/blob/dev/src/webargs/pyramidparser.py

PyramidRequestSchema lets the schema itself declare what part(s) of the
request it reads from rather than the caller having to do so, and
declare it in a way that's decoupled from the tricky details of WebOb:

    class MySchema(PyramidRequestSchema):
        locations = ["form"]

    parsed_params = MySchema(request).parse()

`locations` can also be overridden on a per-call basis if desired.