WIP: Add canvasapi command-line script
This adds a canvasapi command-line script (call it like this:
`tox -qe dev --run-command 'canvasapi conf/development.ini'`) that
allows you to call our Canvas API client (the one from
`lms/services/canvas_api.py`) from the command line in order to easily
experiment with the Canvas API.

This script turned out to be unnecessary: it turns out that a Canvas user
can generate a non-expiring access token for their account that can be
easily passed to curl, httpie, etc. So you can easily experiment with the
Canvas API using those tools. It isn't necessary to have a CLI for the
LMS app's Canvas API client. See:
https://community.canvaslms.com/docs/DOC-16005-42121018197

Nonetheless, I'm committing this work on a branch for posterity.

The script will use an existing, working Canvas API access token from
the LMS app's DB, if there is one. If there's no access token in the DB,
or if its expired (and can't be refreshed) then the script will run the
LMS app locally on a host and port and print out a URL for you to open
in a browser in order to get a new access token. After you click the
[Authorize] button in Canvas in the browser window, the script will
automatically continue using the freshly-authorized access token.

The script requires a Canvas LTI user_id and oauth_consumer_key which
are hard-coded to `***` values, so it won't work unless you fix these
`***` values in the code.

The script is also hard-coded to only call Canvas's list_files API for a
hard-coded course ID.

When the script prints out a URL for you to open in order to authorize
us with Canvas and get a new access token, it assumes that you will log
in to Canvas (or are already logged in to Canvas) as the Canvas user
corresponding to the hard-coded LTI user_id and oauth_consumer_key.