Set `SameSite=None` attribute on feature flag cookie
Feature flag cookies are set in a first-party context from the `/flags`
form but we want them to be send when the LMS app is loaded in an
iframe, in a third-party context, inside the LMS.

Per the recipes at [1], this requires setting `SameSite=None; Secure` on
the cookie.

Note that this change will make this method of setting feature flags
unavailable if serving over plain HTTP in development. In that
environment however the feature can be enabled via other means.

[1] https://web.dev/samesite-cookie-recipes/