Rename js-lms-config back to just js-config
This means that we once again have an `LTILaunchResource.js_config`
property and a matching `js-config` CSS class.

Note that the `js-` prefix is a pattern used to denote CSS classes that
exist for the purpose of enabling JavaScript to find things rather than
for styling, see:

https://github.com/hypothesis/frontend-toolkit/blob/master/docs/css-style-guide.md#classes-used-in-code

I'm not going to change that as it's a pattern used in frontend code
throughout all our apps (and it's a good idea too).

I think the addition of "-lms-" was meant to help distinguish this
config object from the other js-hypothesis-config and
js-rpc-server-config objects, but neither of those exists any more.