Tweak pytest config
Centralise the pytest config in a `pytest.ini` file that applies to both
`make test` and `make functests` instead of repeating command line
options in the `pytest` commands in `tox.ini`.

The new `pytest.ini` fails the tests on warnings which was already being
done for `make test` but not for `make functests` which was printing out
a warning. `make functests` will now fail on warnings too, and the one
warning that it was already printing has been silenced.

The new `pytest.ini` also contains `-q`: less output is nicer when the
tests pass.