Loosen contraints on `item` and `selectedItem` prop types
These can be any type that is compatible with the `renderItem` prop. The
table tests pass a list of strings for the items. There is an API design
issue that it may be ambiguous which item is selected if the items are
primitives (eg. strings) since `selectedItem` may compare equal to
multiple items. If `selectedItem` was an index this couldn't happen.