Patch `propTypes.{element, node}` to work with Preact
The standard implementations of these functions work with React elements
but not Preact elements because they rely on React internals via the
`react-is` package.

Add implementations that work with Preact elements instead and
monkey-patch the standard exports. I'm unsure that monkey-patching is
the best approach, as opposed to just importing the custom prop types
directly, but it does avoid confusion when developers try to use
`propTypes.{node, element}`.