Delete test DB tables before running tests
On master the BDD tests pollutes the test DB such that future runs of the unit
tests fail:

```terminal
$ make backend-tests
(the tests pass)
$ make bddtests
(the tests pass)
$ make backend-tests
(a bunch of assertions fail because there's unexpected data in the test DB)
```

The only way to fix this once its happened is to manually delete your test DB.

This is arguably a problem with the BDD tests: they don't seem to clean up after
themselves, whereas the functests and unit tests do.

But as a fix I've changed the functests and unit tests to delete all the test DB
tables at the beginning of a test run. Ensuring that the test DB is in a clean
state before you start a test run seems more robust than trying to always leave
the DB clean _after_ a test run no matter what happened.