Move debug settings into JSConfig class
`BasicLTILaunchViews` inserts some debug info, used by the Gherkin
tests, into the JavaScript config.

Move this JavaScript config mutation into the `JSConfig` class and make
it unconditional: the debug info is now added to the `JSConfig` for all
requests (that have a `JSConfig`). The less variation of the `JSConfig`
between different requests the better.