Expose the JSConfig instance directly
Expose JSConfig itself directly as context.js_config, rather than only
exposing JSConfig.config.

This is because we're going to be adding new methods to JSConfig for the
views to call, so they'll need access to it.

This means that view code that wants to access the config dict directly
now has to do `context.js_config.config`, but that will eventually be
going away as we phase out direct access to this dict.