Sort JSConfig methods alphabetically
There's no particular order for these methods to go in that makes sense
and, in the final version of JSConfig that I have locally,
alphabetically turned out to work well. In future it will keep similar
methods together because similar naming conventions will be used.

This commit just puts things in alphabetical order ready for future PRs
to slot right in.