Move the "authUrl" JavaScript setting into JSConfig
The `"authUrl"` setting is the URL of the popup window that the
JavaScript code opens when it needs the user to get us a new Canvas API
access token by (re-)authorizing us with Canvas. It's actually a URL of
the LMS app itself, not a Canvas URL, but it redirects the popup window
to the appropriate Canvas URL.

This setting should probably be named something else given that it's
actually specific to the Canvas API.

It also belongs in the `"urls"` JavaScript setting (which is a dict of
all the URLs the JavaScript code needs to know about) not as its own
top-level setting.

But this PR limits itself to moving the `"authUrl"` setting into the
`JSConfig` class.

This change means that `"authUrl"` exists in the JS config all the time
not just some of the time, which does no harm.