Move Canvas submission params into JSConfig
Move the initial Canvas submission params (params that the frontend
sends to the backend's `record_canvas_speedgrader_submission()` endpoint
when a student launches an assignment) into the `JSConfig` class.

This isn't all of the Canvas submission params. There are more elsewhere
in the views, but those will be integrated into `JSConfig` later.