Move via_url_callback, via_url and document_url settings into JSConfig
The way this works now is that `BasicLTILaunchViews` just calls
`JSConfig.add_document_url()` (or `JSConfig.add_canvas_file_id()` in the
case of Canvas Files assignments) and, based on the given document URL
or Canvas file ID, `JSConfig` handles setting the `"via_url_callback"`,
`"via_url"` and Canvas submission settings as appropriate.

The actual `document_url` or `canvas_file_id` comes from a different
place for each view. For example for a URL-configured launch it's the
`"url"` request param, for a DB-configured launch it uses various launch
params to find the document URL in the DB, for a Canvas files assignment
it's a Canvas file ID from the request params and not a document URL at
all, etc.

So it remains the view's job to extract the document URL or Canvas file
ID from the appropriate place for that view, and pass it to
`JSConfig.add_document_url()` or `JSConfig.add_canvas_file_id()`. Given
that one piece of information, `JSConfig` then does everything else.