Centralize and deduplicate service fixtures
This adds fixtures to `tests/unit/conftest.py` for mock services for all
our services. We're now mocking all the services correctly in one place,
rather than mocks for particular services being distributed throughout
the tests and having inconsistencies in fixture naming. This also
removes some duplication where we had multiple fixtures in different
test files for mocking the same service.

Individual tests can of course override and modify these global fixtures
in the usual way.

These are not autouse fixtures because that would cause every one of
these fixtures to be executed, including runnning
`mock.create_autospec()` which is slow, for every single test, and that
slows down a full test run significantly. Instead, modules that need
given services to be mocked declare that using a `pytestmark` attribute.
This is equivalent to having a module-level autouse fixture, but for a
fixture that isn't defined in the module.