Move Canvas focused_user setting into JSConfig
There's also a change in behaviour: it now configures the client to
focus on a user whenever there's a `focused_user` launch param.
Previously it only did this when the LMS was Canvas _and_ there's a
`focused_user` launch param. The is-Canvas check was both unnecessary
and incorrect:

* In practice the focused_user launch param only ever appears in Canvas
  SpeedGrader launches, so the presence of focused_user already tells us
  that we must be in Canvas, it's not necessary to additionally check
  by other means that we are in Canvas

* The user focusing behaviour is not desired in Canvas generally, only
  in Canvas SpeedGrader, so the check should be "is Canvas SpeedGrader?"
  not just "is Canvas?" Just checking for the focused_user param already
  achieves this

This also means that if we ever do add the focused_user param in any
non-Canvas environments it will just work.