Change get_by_assignment() to just return GradingInfos
Change `GradingInfoService.get_by_assignment()` to just return
`GradingInfo` objects instead of `(GradingInfo, HUser)` 2-tuples.

The calling code (in `frontend_app.py`) is now responsible for creating
the `HUser` object corresponding to each `GradingInfo` which it needs to
do in order to create the "student" dicts that it puts into the JS
config.

Overall this ends up making the tests for both `GradingInfoService` and
`frontend_app.py` simpler, and is a better division of work between the
two.